requirements_demo.txt
streamlit==1.28.0
pandas==2.2.3
numpy==1.26.4
pdfplumber==0.7.6
python-docx==0.8.11
tika==2.6.0
python-dotenv==1.0.0
phonenumbers==8.13.18
email_validator==2.0.0
rapidfuzz==2.13.7
joblib==1.3.2
langchain==0.0.315
